Add GitHub action to sync README and CRDs
There are two sets of files that need to be synchronized in different
places.

First is the README. The primary website for the plugin is hosted here
on GitHub pages: https://cert-manager.github.io/aws-privateca-issuer/
and this page is represented by the `index.md` file on the gh-pages
branch. This file should be consistent with the contents of `README.md`
on the master branch to make sure customers see the latest information.

Second is the CRDs (Custom Resource Definitions). We define CRDs in the
config/crd/bases/ directory. However, when we publish charts to helm, we
publish the CRDs located in the charts/aws-pca-issuer/crds directory. In
order to ensure customers are getting the latest custom resources we
need to copy the CRDs in the config directory to the chart directory.

Signed-off-by: Abhishek Mishra <mibhi@amazon.com>