Make CRD sync use pull requests and be bidirectional
Originally to sync CRDs, we would wait for CRD changes to be pushed to
the `config` directory. Once this happened, we would copy these changes
to the `charts` directory and push these changes directly to master.
However there were concerns with this, especially regarding safety
concerns when directly pushing to master, and the possibility that our
repo would block direct pushes to master and require a PR.

Therefore, our new CRD sync workflow will wait for a change to be made
to *either* the `config` or `charts` directory in a *pull-request*. We
then check which set of CRDs were modified. If only CRDs in `config` are
modified, we overwrite the unmodified `charts` directory with these
changes, and vice versa. If both directories have changes, we check
that all CRDs are identical in both directories. If they are not
identical we leave a comment on the PR and the workflow fails.

The reason we use this bidirectional approach and only overwrite
unmodified changes is because of how the update-pull-request action
works. It will checkout the pull request, make and commit local changes,
but then rebase changes using the `-XTheirs` flag. This means any merge
conflicts when rebasing will use the original contribution instead. As a
result, we cannot overwrite any contributor changes, only change
unmodified code. This means we either have to overwrite unmodified CRDs
or check that they're already equal -- we cannot enforce a certain
directory as a source of truth.

Signed-off-by: Abhishek Mishra <mibhi@amazon.com>