Don't sync CRDs for PRs from external forks
The GitHub token provided by GitHub actions does not allow us to add
commits to external forks. Since this workflow depends on making a
commit to sync CRDs during a PR, we are unable to sync CRDs for external
forks.

Instead, we simply check if the CRDs are equal in the charts and config
directories. If they are equal the workflow passes as usual. If they are
different, the workflow will fail and automatically leave  a comment on the PR
asking the contributor to sync the CRDs manually.

Signed-off-by: Abhishek Mishra <mibhi@amazon.com>