Set on-release.yml to trigger on release publishing instead of push
Previously, this workflow triggered on a push with any tag. This was
causing a weird edge case where workflows that were triggered on a
forked branch of the repository with a release cut to it would show up
on the parent repo.

If we change this workflow to trigger instead on the actual publishing
of a release, this behavior should stop.

Signed-off-by: HanleyYin <yhanley@amazon.com>